+++ src/blocks/Tabs/schema.ts
@@ -3,7 +3,7 @@ import {
     LinkProps,
     withTheme,
     BlockBaseProps,
-    TitleProps,
+    BlockHeaderProps,
     MediaProps,
     containerSizesArray,
     sizeNumber,
@@ -44,13 +44,7 @@ export const TabsBlock = {
         required: ['title', 'items'],
         properties: {
             ...BlockBaseProps,
-            title: {
-                oneOf: [{type: 'string'}, TitleProps],
-            },
-            description: {
-                type: 'string',
-                contentType: 'text',
-            },
+            ...BlockHeaderProps,
             tabsColSizes: containerSizesArray.reduce(
                 (acc, size) => ({...acc, [size]: sizeNumber}),
                 {},